UI-Evol/requirements.txt
PyYAML>=6.0
Pillow>=9.0.0
azure-identity>=1.12.0
azure-core>=1.24.0
openai>=1.0.0
tenacity>=8.0.0
numpy>=1.21.0
natsort>=8.0.0 